dont run mvn on a non-maven repo

+++ release.sh
@@ -16,12 +16,17 @@ then
     echo "* pushing to origin"
     git checkout ${RELEASE_VERSION}
     echo "DEPLOY_EXISTING: ${DEPLOY_EXISTING}"
-    if [ -n "${MAVEN_PUSH}" ]
+    if [ -e "pom.xml" ];
     then
-      echo 'deploying existing repo'
-      mvn clean deploy -DperformRelease -DskipTests ${BAMBOO_OPTS}
+      if [ -n "${MAVEN_PUSH}" ]
+      then
+        echo 'deploying existing repo'
+        mvn clean deploy -DperformRelease -DskipTests ${BAMBOO_OPTS}
+      else
+        mvn clean install -DskipTests
+      fi
     else
-      mvn clean install -DskipTests
+      echo "No pom.xml - not building"
     fi
 else
   echo SNAPSHOT_VERSION=${SNAPSHOT_VERSION}
